+++ .gitignore
@@ -43,10 +43,9 @@ yarn-error.log
 
 # dotenv environment variable files
 .env
-.env.development.local
-.env.test.local
-.env.production.local
-.env.local
+.env.development
+.env.test
+.env.production
 
 
 # fastlane
