+++ package.json
@@ -6,7 +6,10 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios --simulator=\"iPhone 14\"",
     "lint": "eslint .",
-    "start": "NODE_ENV=development react-native start",
+    "start": "react-native start",
+    "start:clear": "react-native start --reset-cache",
+    "start:dev": "APP_ENV=development react-native start",
+    "start:prod": "APP_ENV=production react-native start",
     "test": "jest",
     "storybook-generate": "sb-rn-get-stories",
     "storybook-watch": "sb-rn-watcher"
