+++ src/App.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import {API_KEY_PROD, API_KEY_DEV} from '@env';
+import {API_KEY} from '@env';
 
 import Router from './navigation/Router';
 
 import ReduxProvider from './redux/ReduxProvider';
 
 const App: React.FC = () => {
-  console.log(API_KEY_PROD);
-  console.log(API_KEY_DEV);
+  console.log(API_KEY);
+
   return (
     <ReduxProvider>
       <Router />
