+++ src/components/General/Button/Button.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import {TouchableOpacity, Text, StyleSheet} from 'react-native';
+
+export interface ButtonProps {
+  onPress: () => void;
+  text: string;
+}
+
+const Button: React.FC<ButtonProps> = ({onPress, text}) => {
+  return (
+    <TouchableOpacity
+      style={styles.container}
+      onPress={onPress}
+      activeOpacity={0.8}>
+      <Text style={styles.text}>{text}</Text>
+    </TouchableOpacity>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    paddingHorizontal: 16,
+    paddingVertical: 8,
+    backgroundColor: 'purple',
+    borderRadius: 8,
+  },
+  text: {color: 'white'},
+});
+
+export default Button;
