+++ src/types/env.d.tsx
@@ -1,4 +1,3 @@
 declare module '@env' {
-  export const API_KEY_PROD: string;
-  export const API_KEY_DEV: number;
+  export const API_KEY: string;
 }
